criu: Version 1.0 release
The next version is ready. It looks like it's time to make it
be the first major release, so here it is -- the v1.0. It's as
functional as v0.8 but with a lot of bug fixes and improvements.
And it's still :) compatible wrt images and cli options with
older criu-s.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>