irmap: Properly compare stat vs proc devices
The s_dev value we see in irmap is get from proc file and is
raw kernel dev_t. We compare this value to the device get from
stat syscall, which is old-style dev_t (with less bits for minor).

Thus, need to convert kernel dev_t to stat's dev_t for proper
comparison.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>